+++ src/main/java/com/alchemist/service/CommandUtil.java
@@ -1,0 +1,7 @@
+package com.alchemist.service;
+
+public class CommandUtil {
+	public static String [] parseCommand(String command) {
+		return command.split("\\s+");
+	}
+}
